// frontend/src/api/model/PatientProfileDto.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { Allergy } from './Allergy';
import type { EmergencyContact } from './EmergencyContact';

export interface PatientProfileDto {
  userPublicId: string;
  patientPublicId: string;
  firstName: string;
  lastName: string;
  email: string;
  role: string;
  /** @nullable */
  bloodGroup: string | null;
  emergencyContact: null | EmergencyContact;
  /** @nullable */
  allergies: Allergy[] | null;
}